Extract ClassConfig into own file
