+++ lib/vorpal/dsl/config_builder.rb
@@ -96,7 +96,7 @@ module Vorpal
     private
 
     def build_class_config
-      Vorpal::ClassConfig.new(
+      Vorpal::Config::ClassConfig.new(
         domain_class: @domain_class,
         db_class: @class_options[:to] || @defaults_generator.build_db_class(@class_options[:table_name]),
         serializer: @class_options[:serializer] || @defaults_generator.serializer(attributes_with_id),
